+++ tests/sdk/v2/driver/test_ce_runtime.py
@@ -7,6 +7,7 @@ from unittest.mock import DEFAULT, MagicMock, Mock, call, create_autospec
 import pytest
 
 from orquestra.sdk import exceptions
+from orquestra.sdk._base import serde
 from orquestra.sdk._base._driver import _ce_runtime, _client, _exceptions, _models
 from orquestra.sdk._base._testing._example_wfs import (
     my_workflow,
@@ -454,7 +455,7 @@ class TestGetWorkflowRunResultsNonBlocking:
         # Then
         mocked_client.get_workflow_run_results.assert_called_once_with(workflow_run_id)
         mocked_client.get_workflow_run_result.assert_has_calls([call("result_id")])
-        assert results == (1,)
+        assert results == (JSONResult(value="[1]"),)
 
     def test_happy_path_tuple(
         self,
@@ -463,10 +464,14 @@ class TestGetWorkflowRunResultsNonBlocking:
         workflow_run_id: str,
     ):
         # Given
-        mocked_client.get_workflow_run_results.return_value = ["result_id"]
-        # Currently, the result is JSON serialised, which means the tuple information
-        # is discarded
-        mocked_client.get_workflow_run_result.return_value = JSONResult(value="[1, 2]")
+        mocked_client.get_workflow_run_results.return_value = [
+            "result_id",
+            "result_id2",
+        ]
+        mocked_client.get_workflow_run_result.side_effect = [
+            JSONResult(value="1"),
+            JSONResult(value="2"),
+        ]
 
         # When
         results = runtime.get_workflow_run_outputs_non_blocking(workflow_run_id)
@@ -474,7 +479,7 @@ class TestGetWorkflowRunResultsNonBlocking:
         # Then
         mocked_client.get_workflow_run_results.assert_called_once_with(workflow_run_id)
         mocked_client.get_workflow_run_result.assert_has_calls([call("result_id")])
-        assert results == (1, 2)
+        assert results == (JSONResult(value="1"), JSONResult(value="2"))
 
     class TestGetWorkflowRunResultsFailure:
         def test_bad_workflow_run_id(
@@ -643,7 +648,7 @@ class TestGetAvailableOutputs:
     ):
         # Given
         mocked_client.get_workflow_run_artifacts.return_value = {
-            f"{workflow_run_id}@task-inv-1": ["wf-art-1", "wf-art-2"],
+            f"{workflow_run_id}@task-inv-1": ["wf-art-1"],
             f"{workflow_run_id}@task-inv-2": ["wf-art-3"],
         }
         mocked_client.get_workflow_run_artifact.return_value = JSONResult(value="1")
@@ -656,11 +661,11 @@ class TestGetAvailableOutputs:
             workflow_run_id
         )
         mocked_client.get_workflow_run_artifact.assert_has_calls(
-            [call("wf-art-1"), call("wf-art-2"), call("wf-art-3")]
+            [call("wf-art-1"), call("wf-art-3")]
         )
         assert results == {
-            "task-inv-1": (1, 1),
-            "task-inv-2": (1,),
+            "task-inv-1": JSONResult(value="1"),
+            "task-inv-2": JSONResult(value="1"),
         }
 
     class TestGetWorkflowRunArtifactsFailure:
@@ -741,7 +746,7 @@ class TestGetAvailableOutputs:
         @pytest.fixture
         def mocked_client(self, mocked_client: MagicMock, workflow_run_id):
             mocked_client.get_workflow_run_artifacts.return_value = {
-                f"{workflow_run_id}@task-inv-1": ["wf-art-1", "wf-art-2"],
+                f"{workflow_run_id}@task-inv-1": ["wf-art-1"],
                 f"{workflow_run_id}@task-inv-2": ["wf-art-3"],
             }
             return mocked_client
@@ -763,7 +768,7 @@ class TestGetAvailableOutputs:
                 workflow_run_id
             )
             mocked_client.get_workflow_run_artifact.assert_has_calls(
-                [call("wf-art-1"), call("wf-art-2"), call("wf-art-3")]
+                [call("wf-art-1"), call("wf-art-3")]
             )
             assert results == {}
 
@@ -777,7 +782,6 @@ class TestGetAvailableOutputs:
             mocked_client.get_workflow_run_artifact.return_value = JSONResult(value="1")
             mocked_client.get_workflow_run_artifact.side_effect = (
                 DEFAULT,
-                DEFAULT,
                 Exception,
             )
 
@@ -789,9 +793,9 @@ class TestGetAvailableOutputs:
                 workflow_run_id
             )
             mocked_client.get_workflow_run_artifact.assert_has_calls(
-                [call("wf-art-1"), call("wf-art-2"), call("wf-art-3")]
+                [call("wf-art-1"), call("wf-art-3")]
             )
-            assert results == {"task-inv-1": (1, 1)}
+            assert results == {"task-inv-1": JSONResult(value="1")}
 
         def test_continues_after_failure(
             self,
@@ -804,7 +808,6 @@ class TestGetAvailableOutputs:
             mocked_client.get_workflow_run_artifact.side_effect = (
                 Exception,
                 DEFAULT,
-                DEFAULT,
             )
 
             # When
@@ -815,11 +818,10 @@ class TestGetAvailableOutputs:
                 workflow_run_id
             )
             mocked_client.get_workflow_run_artifact.assert_has_calls(
-                [call("wf-art-1"), call("wf-art-2"), call("wf-art-3")]
+                [call("wf-art-1"), call("wf-art-3")]
             )
             assert results == {
-                "task-inv-1": (1,),
-                "task-inv-2": (1,),
+                "task-inv-2": JSONResult(value="1"),
             }
 
         def test_unknown_http(
@@ -841,7 +843,7 @@ class TestGetAvailableOutputs:
                 workflow_run_id
             )
             mocked_client.get_workflow_run_artifact.assert_has_calls(
-                [call("wf-art-1"), call("wf-art-2"), call("wf-art-3")]
+                [call("wf-art-1"), call("wf-art-3")]
             )
             assert results == {}
 
@@ -866,7 +868,7 @@ class TestGetAvailableOutputs:
                 workflow_run_id
             )
             mocked_client.get_workflow_run_artifact.assert_has_calls(
-                [call("wf-art-1"), call("wf-art-2"), call("wf-art-3")]
+                [call("wf-art-1"), call("wf-art-3")]
             )
             assert results == {}